legacy/requirements.txt
# API dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2

# YouTube transcript dependencies
google-api-python-client>=2.108.0
youtube-transcript-api>=0.6.1
scrapetube>=0.1.0

# Knowledge graph dependencies
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-experimental>=0.0.49
openai>=1.3.0
tiktoken>=0.5.1

# Utility dependencies
python-dotenv>=1.0.0
isodate>=0.6.1
tqdm>=4.66.1

annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
decorator==5.2.1
distro==1.9.0
executing==2.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpickle==4.1.0
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.13
narwhals==1.40.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
parso==0.8.4
pexpect==4.9.0
plotly==6.1.1
preshed==3.0.9
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic_core==2.33.2
Pygments==2.19.1
pyvis==0.3.2
requests==2.32.3
rich==14.0.0
setuptools==80.8.0
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
thinc==8.3.6
traitlets==5.14.3
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wheel==0.45.1
wrapt==1.17.2
